requirements.txt
torch
numpy
pandas
opencv-python
pillow
torchvision
scikit-learn
scikit-image
joblib